+++ src/CarbonFromTimestampCast.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace LittleGreenMan\Earhart;
+
+use Carbon\Carbon;
+use Spatie\LaravelData\Casts\Cast;
+use Spatie\LaravelData\Support\Creation\CreationContext;
+use Spatie\LaravelData\Support\DataProperty;
+
+class CarbonFromTimestampCast implements Cast
+{
+    public function cast(DataProperty $property, mixed $value, array $properties, CreationContext $context): mixed
+    {
+        return Carbon::createFromTimestamp($value);
+    }
+}