+++ src/Controllers/AuthOrgCreateController.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace LittleGreenMan\Earhart\Controllers;
+
+use Illuminate\Http\Request;
+
+class AuthOrgCreateController
+{
+    /**
+     * Handle the incoming request.
+     */
+    public function __invoke(Request $request)
+    {
+        return redirect(config('services.propelauth.auth_url') . '/create_org');
+    }
+}