+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `earhart` will be documented in this file.
 
+# v1.2.0
+- Added an initial API library to support getting Organisations, an Organisation and Users in an Organisation.
+- Added routes:
+  - Org Members /org/members/:ordId
+  - Org Settings /org/settings/:ordId
+  - Create Org /create_org
+  - Account Settings /account/settings/:orgId
+
 # v1.1.0
 - Added AuthAccountController to provide redirect to PropelAuth account manager.